Rename file feature (#147)
* Rename to config_parsing

* Add workflows for devel
