+++ .github/workflows/main.yml
@@ -4,9 +4,11 @@ on:
   push:
     branches:
     - master
+    - devel
   pull_request:
     branches:
     - master
+    - devel
 
 jobs:
   lint:
