+++ src/append/rolling_file/policy/compound/trigger/mod.rs
@@ -5,8 +5,8 @@ use std::fmt;
 use failure::Error;
 
 use crate::append::rolling_file::LogFile;
-#[cfg(feature = "file")]
-use crate::file::Deserializable;
+#[cfg(feature = "config_parsing")]
+use crate::config_parsing::Deserializable;
 
 #[cfg(feature = "size_trigger")]
 pub mod size;
@@ -17,7 +17,7 @@ pub trait Trigger: fmt::Debug + Send + Sync + 'static {
     fn trigger(&self, file: &LogFile) -> Result<bool, Error>;
 }
 
-#[cfg(feature = "file")]
+#[cfg(feature = "config_parsing")]
 impl Deserializable for dyn Trigger {
     fn name() -> &'static str {
         "trigger"
