+++ src/lib.rs
@@ -134,7 +134,7 @@
 //! Add the following in your application initialization.
 //!
 //! ```no_run
-//! # #[cfg(feature = "file")]
+//! # #[cfg(feature = "config_parsing")]
 //! # fn f() {
 //! log4rs::init_file("log4rs.yml", Default::default()).unwrap();
 //! # }
@@ -189,11 +189,9 @@
 use arc_swap::ArcSwap;
 use fnv::FnvHasher;
 use log::{Level, LevelFilter, Metadata, Record, SetLoggerError};
-use std::{
-    cmp, collections::HashMap, hash::BuildHasherDefault, io, io::prelude::*, sync::Arc,
-};
+use std::{cmp, collections::HashMap, hash::BuildHasherDefault, io, io::prelude::*, sync::Arc};
 
-#[cfg(feature = "file")]
+#[cfg(feature = "config_parsing")]
 pub use crate::priv_file::{init_file, load_config_file, FormatError};
 
 use crate::{append::Append, config::Config, filter::Filter};
@@ -201,11 +199,13 @@ use crate::{append::Append, config::Config, filter::Filter};
 pub mod append;
 pub mod config;
 pub mod encode;
-#[cfg(feature = "file")]
-pub mod file;
 pub mod filter;
-#[cfg(feature = "file")]
+
+#[cfg(feature = "config_parsing")]
+pub mod config_parsing;
+#[cfg(feature = "config_parsing")]
 mod priv_file;
+
 #[cfg(feature = "console_writer")]
 mod priv_io;
